requirements.txt
aws-cdk-lib==2.92.0
constructs==10.2.69
